Remove Home page, disable all buttons on Navbar, and rename to QuestionNavbar
