+++ frontend/src/App.js
@@ -1,15 +1,13 @@
 import { Route, Routes } from "react-router-dom";
-import logo from './styles/logo.svg';
 import './styles/App.css';
-import { Home, Questions } from './pages';
+import { Questions } from './pages';
 
 function App() {
   return (
     <div className="App">
 
       <Routes>
-        <Route path="/" element={<Home />} />
-        <Route path="/questions" element={<Questions />} />
+        <Route path="/" element={<Questions />} />
       </Routes>
     </div>
   );
