+++ frontend/src/pages/Questions.jsx
@@ -1,4 +1,4 @@
-import Navbar from "../components/Navbar";
+import QuestionNavbar from "../components/QuestionNavbar";
 import QuestionTable from '../components/QuestionTable'
 import AddQuestionButton from '../components/AddQuestionButtons'
 import '../styles/questions.css';
@@ -6,7 +6,7 @@ import '../styles/questions.css';
 const Questions = () => {
     return (
         <div>
-            <Navbar />
+            <QuestionNavbar />
             <h1>Questions</h1>
             <p className="description">View all the questions stored in database.</p>
             <div className="question-table-container">
